+++ app/lizard-nxt.js
@@ -47,7 +47,8 @@ angular.module("lizard-nxt", [
   'lodash',
   'ui.bootstrap',
   'lizard-http-throttler', // Add this $http interceptor befor the loading-bar.
-  'angular-loading-bar'
+  'angular-loading-bar',
+  'lizard-boostrap'
 ])
 
 // Decorator for ngExceptionHandler to log exceptions to sentry
@@ -62,8 +63,11 @@ angular.module("lizard-nxt", [
   });
 })
 
+// The internet says it is better to set debugInfoEnabled to false in production
+// but in dev it is still handy to have it set to true when compiling. No access
+// to anything from the server yet, so checking window.location.
 .config(['$compileProvider', function ($compileProvider) {
-  $compileProvider.debugInfoEnabled(false);
+  $compileProvider.debugInfoEnabled(window.location.host === 'localhost:9000');
 }])
 
 /**
@@ -96,13 +100,6 @@ angular.module("lizard-nxt", [
 ])
 
 /**
- * @name user
- * @memberOf app
- * @description User and auth stuff
- */
-.constant('user', window.user)
-
-/**
  * @name notie
  * @memberOf app
  * @description Notification service
@@ -110,29 +107,61 @@ angular.module("lizard-nxt", [
 .constant('notie', window.notie)
 
 /**
- * @name versioning
- * @memberOf app
- * @description User and auth stuff
- */
-.constant('versioning', window.versioning)
-
-/**
  * @name production backend
  * @memberOf app
  * @description subdomain of production backend.
  */
-.constant('backendDomain', 'https://demo.lizard.net')
+.constant('backendDomain', 'https://demo.lizard.net');
 
-/**
- * @name locale
- * @memberOf app
- * @description Portal's default locale.
- */
-.constant('defaultLocale', window.locale)
+angular.module('lizard-boostrap', ['favourites'])
+.run([
+  '$http', 'UrlService', 'FavouritesService', 'user', 'version', 'debug',
+  function ($http, UrlService, FavouritesService, user, version, debug) {
 
-/**
- * @name temporalBounds
- * @memberOf app
- * @description Portal's default temporal bounds.
- */
-.constant('temporalBounds', window.temporalBounds);
+    var urlState = UrlService.getState();
+
+    var showErrorModal = function () {
+      var overlay = document.getElementById('dark-overlay');
+      overlay.style.display = 'inline';
+      throw new Error('No lizard/bootstrap.json lizard is down or malfunctioning');
+    };
+
+    var getBootstrap = function (applyState) {
+      $http.get('bootstrap/lizard/', {})
+      .then(
+        function (response) {
+          var bootstrap = response.data;
+          _.merge(user, bootstrap.user);
+          _.merge(version, bootstrap.version);
+          if (applyState) {
+            FavouritesService.applyFavourite(bootstrap.state);
+            FavouritesService.applyFavourite(urlState);
+          }
+        },
+        function (response) {
+          showErrorModal();
+        }
+      );
+    };
+
+    var urlFavourite = UrlService.getFavourite();
+
+    if (urlFavourite) {
+      FavouritesService.getFavourite(
+        urlFavourite,
+        function (favourite, getResponseHeaders) {
+          getBootstrap(false);
+          FavouritesService.applyFavourite(favourite);
+        },
+        function () {
+          // Show error facing message
+          getBootstrap(true);
+        }
+      );
+    }
+    else {
+      getBootstrap(true);
+    }
+
+  }
+]);