+++ CHANGES.rst
@@ -24,6 +24,17 @@ Release 3.2.5 (2016-7-13)
 Release 3.2.4 (2016-6-20)
 ---------------------
 
+- Remove LayerGroups, use dataLayers and mapLayers with layer directives in
+  data-menu.
+
+- Use bootstrap/lizard in json in favor of lizard-bs.js.
+
+- Use favouriteService.applyFavourite for bootstrapping and url state.
+
+
+Release 3.2.4 (2016-6-20)
+---------------------
+
 - Fix no attributes for nested assets